Module lol_html::errors

The errors that can be produced by the crate’s API.

Structs

MemoryLimitExceededError

An error that occures when rewriter exceedes the memory limit specified in the MemorySettings.

ParsingAmbiguityError

An error that occurs when HTML parser runs into an ambigious state in the strict mode.

Enums

AttributeNameError

An error that occurs when invalid value is provided for the attribute name.

CommentTextError

An error that occurs when invalid value is provided for the HTML comment text.

RewritingError

A compound error type that can be returned by write and end methods of the rewriter.

SelectorError

A CSS selector parsing error.

TagNameError

An error that occurs when invalid value is provided for the tag name.