Module lol_html::html_content

HTML content descriptors that can be produced and modified by a rewriter.

Structs

Attribute
An attribute of an Element.

Comment
An HTML comment rewritable unit.

Doctype
A document type declaration preamble.

DocumentEnd
A rewritable unit that represents the end of the document.

Element
An HTML element rewritable unit.

TextChunk
An HTML text node chunk.

Enums

ContentType
The type of inserted content.

TextType
A type of parsed text.

Traits

UserData
Data that can be attached to a rewritable unit by a user and shared between content handler invocations.